requirements.txt
dbt-core>=1.10.0
dbt-postgres>=1.9.0
psycopg2-binary>=2.9.0
anthropic>=0.30.0
